upai: drm: Add two new fourcc codes for packed yuv444 video formats
The Xilinx Video Mixer IP support dma of memory with video information
stored in yuv444 as packed video data.  There are no fourcc codes
in the user api for DRM to support communicating this capability to
user space applications.   This patch serves as a proposal for these
new codes in support of our IP.

Signed-off-by: Jeffrey Mouroux <jmouroux@xilinx.com>