drm: xilinx: Fix layer update performance issues in mixer
Some changes to mixer layers require a layer to be disabled prior
to applying the change.  The layer enable control register is only
updated on frame boundaries necessitating a delay between a call
to disable and then later enable.  This change ensures that
the mdelay() calls which were consuming CPU are replaced with
msleep() calls.  In addition, a layer is only disabled if the
call to SET_PLANE from user space updates height or width.  Changing
crtc x/y only will not result in a disable or any required
sleep.

Signed-off-by: Jeffrey Mouroux <jmouroux@xilinx.com>